airzone: add changes suggested by tkdrob
